requirements.txt
attrs==21.4.0
colorama==0.4.4
docopt==0.6.2
flake8==4.0.1
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pyserial==3.5
pytest==6.2.5
pytest-watch==4.2.0
toml==0.10.2
watchdog==2.1.6
